added database v3

+++ water_management/src/main/java/com/punjab/water/management/water_management/controller/TehsilController.java
@@ -22,7 +22,7 @@ public class TehsilController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Tehsil> getTehsilById(@PathVariable Long id) {
+    public ResponseEntity<Tehsil> getTehsilById(@PathVariable Integer id) {
         Optional<Tehsil> tehsil = tehsilService.findById(id);
         return tehsil.map(ResponseEntity::ok).orElse(ResponseEntity.notFound().build());
     }
@@ -34,7 +34,7 @@ public class TehsilController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Tehsil> updateTehsil(@PathVariable Long id, @RequestBody Tehsil tehsil) {
+    public ResponseEntity<Tehsil> updateTehsil(@PathVariable Integer id, @RequestBody Tehsil tehsil) {
         Optional<Tehsil> existingTehsil = tehsilService.findById(id);
         if (existingTehsil.isPresent()) {
             // Set the ID from path variable to ensure we update the correct record
@@ -46,7 +46,7 @@ public class TehsilController {
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteTehsil(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteTehsil(@PathVariable Integer id) {
         Optional<Tehsil> tehsil = tehsilService.findById(id);
         if (tehsil.isPresent()) {
             tehsilService.deleteById(id);
